+++ src/constructs/dlz-bastion-host/dlz-bastion-host.ts
@@ -2,7 +2,7 @@ import * as ec2 from 'aws-cdk-lib/aws-ec2';
 import * as iam from 'aws-cdk-lib/aws-iam';
 import * as ssm from 'aws-cdk-lib/aws-ssm';
 import { Construct } from 'constructs';
-import { BastionHost } from '../../data-landing-zone';
+import { BastionHost } from '../../data-landing-zone-types';
 import { SSM_PARAMETERS_DLZ } from '../../stacks/organization/constants';
 
 export interface DlzBastionHostProps {
