+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './data-landing-zone';
+export * from './data-landing-zone-types';
 export * from './data-landing-zone-client';
 export * from './stacks';
 export * from './constructs';
