+++ src/stacks/organization/workloads/network-connections-phase-2-stack/regional-stack.ts
@@ -1,7 +1,8 @@
 import { Construct } from 'constructs';
 import { Shared } from './shared';
 import { DlzStack } from '../../../../constructs/index';
-import { DataLandingZoneProps, WorkloadAccountProps } from '../../../../data-landing-zone';
+
+import { DataLandingZoneProps, WorkloadAccountProps } from '../../../../data-landing-zone-types';
 
 export class WorkloadRegionalNetworkConnectionsPhase2Stack extends DlzStack {
