+++ src/stacks/organization/workloads/network-connections-phase-2-stack/shared.ts
@@ -1,7 +1,7 @@
 import * as ec2 from 'aws-cdk-lib/aws-ec2';
 import * as ssm from 'aws-cdk-lib/aws-ssm';
 import { DlzSsmReader, DlzStack, NetworkEntityVpc } from '../../../../constructs/index';
-import { DataLandingZoneProps, GlobalVariables } from '../../../../data-landing-zone';
+import { DataLandingZoneProps, GlobalVariables } from '../../../../data-landing-zone-types';
 import { Logger } from '../../../../lib/logger';
 import { SSM_PARAMETERS_DLZ } from '../../constants';
